+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# v2.0.10
+
+### Security
+- Fix a bug where blocked users were still able to message modmail.
+
+# v2.0.9
+
+### Added 
+- Support for custom blocked emoji and sent emoji
+- Use the `config set blocked_emoji` or `sent_emoji` commands.
+
+### Quick Fix
+- Support multiple image and file attachments in one message
+- This is only possible on mobile so its good to handle it in code.
 
 # v2.0.8